docs(tailwind): move Tailwind section under Systems
