+++ docs/Systems/LayoutAndSpacing/LayoutAndSpacing.stories.mdx
@@ -24,7 +24,7 @@ Engineers need to use `rem` when spacing for accessibility reasons. However, Fig
 
 ## Implementation
 
-Kaizen's spacing tokens are easily available through our [Tailwind package](/story/tailwind-overview--page).
+Kaizen's spacing tokens are easily available through our [Tailwind package](/story/systems-tailwind-overview--page).
 
 For example, if a design specifies a `div` with a margin-right of `12px`, the spacing token `12` can be applied to Tailwind's `mr-` utility class.
