+++ docs/Systems/Tailwind/ClassnameReferences/BorderColor.stories.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
+import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
+import { CATEGORIES } from "../../../../storybook/constants"
 import { flattenEntries } from "../helpers/flatten-entries"
 import styles from "./styles.module.scss"
 
@@ -13,7 +14,7 @@ const classEntries = flattenEntries(
 )
 
 export default {
-  title: "Tailwind/Classname References/Borders/Border Color",
+  title: `${CATEGORIES.tailwind}/Classname References/Borders/Border Color`,
   parameters: {
     docs: {
       description: {
