+++ docs/Systems/Tailwind/ClassnameReferences/LeadingHeight.stories.tsx
@@ -2,14 +2,17 @@ import React from "react"
 import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
+import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
+import { CATEGORIES } from "../../../../storybook/constants"
+import styles from "./styles.module.scss"
 
-const prefix = "font-"
+const prefix = "leading-"
 const classKeyVal: string[][] = Object.entries(
-  kaizenTailwindTheme?.fontFamily || []
+  kaizenTailwindTheme?.lineHeight || []
 )
+
 export default {
-  title: "Tailwind/Classname References/Typography/Font Family",
+  title: `${CATEGORIES.tailwind}/Classname References/Typography/Line Height`,
   parameters: {
     docs: {
       description: {
@@ -26,15 +29,20 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
     <StoryWrapper isReversed={isReversed}>
       <StoryWrapper.RowHeader headings={["Class", "Properties", "Example"]} />
       {classKeyVal.map((presetData, index) => {
-        const [fontFamilyClassName, fontFamilyValue] = presetData
+        const [leadingHeightClassName, leadingHeightValue] = presetData
 
         return (
           <React.Fragment key={index}>
             <Divider variant="canvas" />
             <StoryWrapper.Row rowTitle="">
-              <p>font-{fontFamilyClassName}</p>
-              <p>{fontFamilyValue}</p>
-              <p style={{ fontFamily: fontFamilyValue[0] }}>Aa</p>
+              <p>leading-{leadingHeightClassName}</p>
+              <p>{leadingHeightValue}</p>
+              <p
+                style={{ lineHeight: leadingHeightValue }}
+                className={styles.wrappedText}
+              >
+                Tutant Meenage Neetle Teetles
+              </p>
             </StoryWrapper.Row>
           </React.Fragment>
         )
@@ -44,5 +52,5 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
 )
 
 export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Font Family"
+StickerSheetDefault.storyName = "Line Height"
 StickerSheetDefault.parameters = { chromatic: { disable: false } }
