+++ docs/Systems/Tailwind/ClassnameReferences/TextColor.stories.tsx
@@ -2,14 +2,15 @@ import React from "react"
 import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
+import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
+import { CATEGORIES } from "../../../../storybook/constants"
 import { flattenEntries } from "../helpers/flatten-entries"
 import styles from "./styles.module.scss"
 
 const prefix = "text-"
 const classEntries = flattenEntries(prefix, kaizenTailwindTheme?.colors || {})
 export default {
-  title: "Tailwind/Classname References/Typography/Text Color",
+  title: `${CATEGORIES.tailwind}/Classname References/Typography/Text Color`,
   parameters: {
     docs: {
       description: {
