+++ docs/Systems/Tailwind/Contributing.stories.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { CATEGORIES } from "../../storybook/constants"
+import { CATEGORIES } from "../../../storybook/constants"
 
 <Meta title={`${CATEGORIES.tailwind}/Contributing to this package`} />
 
@@ -18,7 +18,7 @@ Any time you update the Tailwind presets, make sure you run `yarn build` in the 
 
 **Intellisense not working within the class attributes**
 
-For configuring IntelliSense refer to the [CONFIGURATION.md](CONFIGURATION.md)
+For configuring IntelliSense refer to the [CONFIGURATION.md](/story/systems-tailwind-preset-configuration--page)
 
 Check the output in your terminal output for TailwindCSS IntelliSense.
