+++ docs/Systems/Tailwind/GettingStarted.stories.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { CATEGORIES } from "../../storybook/constants"
+import { CATEGORIES } from "../../../storybook/constants"
 import TailwindPlay from "./assets/tailwind-play.gif"
 import CssGeneration from "./assets/css-generation.png"
 import InlineFold from "./assets/inline-fold.gif"
@@ -50,7 +50,7 @@ module.exports = {
 }
 ```
 
-For preset configuration options, see our [configuration docs](/story/tailwind-overview--page)
+For preset configuration options, see our [configuration docs](/story/systems-tailwind-preset-configuration--page)
 
 ## 3. Add the Tailwind Directives
