+++ docs/Systems/Tailwind/Overview.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs';
-import { CATEGORIES } from "../../storybook/constants"
+import { CATEGORIES } from "../../../storybook/constants"
 import AnatomyDiagram from './assets/tailwind-anatomy.jpg';
 
 <Meta title={`${CATEGORIES.tailwind}/Overview`} />
