+++ docs/Systems/Tailwind/PresetConfiguration.stories.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { CATEGORIES } from "../../storybook/constants"
+import { CATEGORIES } from "../../../storybook/constants"
 
 <Meta title={`${CATEGORIES.tailwind}/Preset configuration`} />
