+++ docs/Systems/Tailwind/WorkingWithTailwind.stories.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { CATEGORIES } from "../../storybook/constants"
+import { CATEGORIES } from "../../../storybook/constants"
 import { Button } from "@kaizen/button"
 import { Heading } from "@kaizen/typography"
