+++ docs/Systems/Tailwind/assets/css-generation.png
