+++ docs/Systems/Tailwind/assets/inline-fold.gif
