+++ docs/Systems/Tailwind/assets/tailwind-anatomy.jpg
